Create QueueCallers object within the middleware
In the process of writing unit tests with redis, we ran into a race
condition with test repository.  So, for now, we have moved to nosetests
and added a Redis fixture to check and make sure our keys have been
deleted.  We'll wait up to 5 seconds for each test.  Not the best
situation, but I don't want to get bogged down with this issue right
now.

Change-Id: Ie7b58cc9001da377b5cb4ff466f82cd68205ffca
Signed-off-by: Paul Belanger <paul.belanger@polybeacon.com>