Create Asterisk application
Here we have a basic example of how AGI can be used with stripe.  This
is a simple demo for the purposes of Astricon.  Still missing is the
actually joining caller to agent logic.

Currently, this application will answer the call and play the position
to the caller every 30 seconds.

Change-Id: I7807b4b75314a925be77097202ef29db6ea73958
Signed-off-by: Paul Belanger <paul.belanger@polybeacon.com>